+++ CHANGELOG.md
@@ -1,9 +1,13 @@
 Change Log
 ==========
 
+## 0.4.0 (2021-07-14)
+* ROS development environment in a Docker container on top of TI Edge AI 0.5 base image
+* Proxy settings are provided for building and running the Docker image behind a proxy network (TI network as an example)
+
 ## 0.3.0 (2021-04-15)
 
-* ROS development environment in a Docker container op top of Processor SDK 7.3.0 pre-built image
+* ROS development environment in a Docker container on top of Processor SDK 7.3.0 pre-built image
 * Enhanced the stereo vision demo application: added point-cloud generation
 * Updated the semantic segmentation demo application: migrated to open-source deep-learning runtime (TVM + Neo-AI-DLR)
 * Added a new demo application: 3D obstacle detection accelerated on deep-learning core (C7/MMA) and hardware accelerators (SDE, LDC, MSC)